Split up whos on first and third base :)
